Added messages for the navigation
Will now give a message telling the user what room they've selected, and when they've reached it.
Currently it's counting the number of corners to the destination, so it may tell you you've arrived even though it's still several meters away. Though the user should be able to find a destination they just have to walk in a straight line to.